general OpenEffectsBox_v2 program ready for beta
generalOpenEffectsBox_v2.ino: instantiates effects according to diagram. Most blocks working, with the exception of things that need restarting periodically, in particular the tonesweeps. Pretty much everything I need to set up a sound is there and working. Next will be assembling the global vars into a state vector, which will enable the creation of presets.