More required modules are implemented
Pedal (connected to the center two stereo jacks in the middle of the jack panel) and Relay (which route the input L and R channels around or through the internal electronics); also the LED strip functionality. Last bit of hardware is the OLED screen - this may in fact be a higher-level function more related to the firmware than the hardware.